Move init code of CndActor from factory fn into actor_test
