+++ api_tests/src/actor_test.ts
@@ -1,8 +1,21 @@
 import { ActorName, Actors } from "./actors";
 import pTimeout from "p-timeout";
-import { HarnessGlobal } from "./environment";
+import { HarnessGlobal, LedgerConfig } from "./environment";
 import { CndActor } from "./actors/cnd_actor";
 import ProvidesCallback = jest.ProvidesCallback;
+import { Logger } from "log4js";
+import { BitcoindWallet, BitcoinWallet } from "./wallets/bitcoin";
+import {
+    newBitcoinStubWallet,
+    newEthereumStubWallet,
+    newLightningStubWallet,
+    Wallets,
+} from "./wallets";
+import { EthereumWallet, Web3EthereumWallet } from "./wallets/ethereum";
+import { LightningWallet } from "./wallets/lightning";
+import { E2ETestActorConfig } from "./config";
+import { merge } from "lodash";
+import { CndInstance } from "./environment/cnd_instance";
 
 declare var global: HarnessGlobal;
 
@@ -22,44 +35,6 @@ export function twoActorTest(
     testFn: (actors: Actors) => Promise<void>
 ): ProvidesCallback {
     return nActorTest(["Alice", "Bob"], testFn);
-}
-
-async function createActors(
-    testName: string,
-    actorNames: ActorName[]
-): Promise<Actors> {
-    const actorsMap = new Map<string, CndActor>();
-
-    const listPromises: Promise<CndActor>[] = [];
-    for (const name of actorNames) {
-        const cndLogFile = global.getLogFile([testName, `cnd-${name}.log`]);
-        const actorLogger = global.getLogger([testName, name]);
-
-        listPromises.push(
-            CndActor.newInstance(
-                name,
-                global.ledgerConfigs,
-                global.cargoTargetDir,
-                cndLogFile,
-                actorLogger,
-                global.cndConfigOverrides,
-                global.gethLockDir,
-                global.lndWallets
-            )
-        );
-    }
-    const createdActors = await Promise.all(listPromises);
-    for (const actor of createdActors) {
-        actorsMap.set(actor.name, actor);
-    }
-
-    const actors = new Actors(actorsMap);
-
-    for (const name of actorNames) {
-        actorsMap.get(name).actors = actors;
-    }
-
-    return actors;
 }
 
 /*
@@ -96,4 +71,110 @@ function nActorTest(
         }
         done();
     };
+}
+
+async function createActors(
+    testName: string,
+    actorNames: ActorName[]
+): Promise<Actors> {
+    const actorsMap = new Map<string, CndActor>();
+
+    const listPromises: Promise<CndActor>[] = [];
+    for (const name of actorNames) {
+        listPromises.push(newCndActor(name, testName));
+    }
+    const createdActors = await Promise.all(listPromises);
+    for (const actor of createdActors) {
+        actorsMap.set(actor.name, actor);
+    }
+
+    const actors = new Actors(actorsMap);
+
+    for (const name of actorNames) {
+        actorsMap.get(name).actors = actors;
+    }
+
+    return actors;
+}
+
+async function newCndActor(name: ActorName, testName: string) {
+    const ledgerConfig = global.ledgerConfigs;
+    const logger = global.getLogger([testName, name]);
+
+    const actorConfig = await E2ETestActorConfig.for(name);
+    const generatedConfig = actorConfig.generateCndConfigFile(ledgerConfig);
+    const finalConfig = merge(generatedConfig, global.cndConfigOverrides);
+    const cndLogFile = global.getLogFile([testName, `cnd-${name}.log`]);
+
+    logger.info(
+        "Created new CndActor with config %s",
+        JSON.stringify(finalConfig)
+    );
+
+    const cndInstance = new CndInstance(
+        global.cargoTargetDir,
+        cndLogFile,
+        logger,
+        finalConfig
+    );
+    const cndStarting = cndInstance.start();
+
+    const bitcoinWallet = newBitcoinWallet(ledgerConfig, logger);
+    const ethereumWallet = newEthereumWallet(
+        ledgerConfig,
+        global.gethLockDir,
+        logger
+    );
+
+    // Await all of the Promises that we started. In JS, Promises are eager and hence already started evaluating. This is an attempt to improve the startup performance of an actor.
+    const wallets = new Wallets({
+        bitcoin: await bitcoinWallet,
+        ethereum: await ethereumWallet,
+        lightning: newLightningWallet(global.lndWallets, name, logger),
+    });
+    await cndStarting;
+
+    return new CndActor(logger, cndInstance, wallets, name);
+}
+
+async function newBitcoinWallet(
+    ledgerConfig: LedgerConfig,
+    logger: Logger
+): Promise<BitcoinWallet> {
+    const bitcoinConfig = ledgerConfig.bitcoin;
+    return bitcoinConfig
+        ? BitcoindWallet.newInstance(bitcoinConfig, logger)
+        : Promise.resolve(newBitcoinStubWallet(logger));
+}
+
+async function newEthereumWallet(
+    ledgerConfig: LedgerConfig,
+    ethereumLockDir: string,
+    logger: Logger
+): Promise<EthereumWallet> {
+    const ethereumConfig = ledgerConfig.ethereum;
+    return ethereumConfig
+        ? Web3EthereumWallet.newInstance(
+              ethereumConfig.dev_account_key,
+              ethereumConfig.rpc_url,
+              logger,
+              ethereumLockDir,
+              ethereumConfig.chain_id
+          )
+        : Promise.resolve(newEthereumStubWallet(logger));
+}
+
+function newLightningWallet(
+    lightningWallets: { alice?: LightningWallet; bob?: LightningWallet },
+    actor: ActorName,
+    logger: Logger
+): LightningWallet {
+    switch (actor) {
+        case "Alice": {
+            return lightningWallets.alice || newLightningStubWallet(logger);
+        }
+        case "Bob": {
+            return lightningWallets.bob || newLightningStubWallet(logger);
+        }
+    }
 }
